pyproject.toml
[project]
name = "mindweave-v1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.3",
    "matplotlib>=3.10.8",
    "openai>=2.30.0",
    "openenv>=0.1.13",
    "openenv-core>=0.2.3",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "torch>=2.11.0",
    "uvicorn>=0.44.0",
]

[project.entry-points."openenv.envs"]
mindweave_v1 = "mindweave_env.server.environment2:MindweaveEnvironment"

[project.scripts]
server = "mindweave_env.server.app:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["mindweave_env*"]
exclude = ["logs*", "models*"]